organize components for editUserProfile
